migrate hook has been redesigned
